feat: kill lowdb code
